requirements.txt
mysql-connector-python~=8.3.0
pygame~=2.5.2